requirements.txt
Cython==0.29.28
fastapi==0.85.0
mxnet==1.6.0
numpy==1.21.5
opencv_contrib_python==4.4.0.46
Pillow==9.2.0
pydantic==1.10.2
scikit_image==0.16.2
setuptools==58.0.4
uvicorn==0.17.6
gunicorn==20.1.0
